+++ Лабораторная 6/Задача 6/Program.cs
@@ -10,8 +10,8 @@ namespace Задача_6
     {
         static void Main(string[] args)
         {
-            double A = double.Parse(Console.ReadLine());
-            double B = double.Parse(Console.ReadLine());
+            int A = int.Parse(Console.ReadLine());
+            int B = int.Parse(Console.ReadLine());
             if(A>B){
             while (B<=A)
             {
